+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The project uses semantic versioning (see [SemVer](https://semver.org)).
 
 ## [Unreleased]
 
+### Fixed
+
+- !setterrain docs no longer mention the wrong command (!settime)
+
 ## v0.47.0 - 2024-04-04
 
 ### Added